requirements.txt
# ═══════════════════════════════════════════
#  Vibe ML — Python Dependencies (v2)
# ═══════════════════════════════════════════

# ── Web Framework ──
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# ── Database ──
sqlalchemy==2.0.36
aiosqlite==0.20.0

# ── Auth ──
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# ── ML Pipeline ──
pandas==2.2.3
numpy==1.26.4
scipy==1.14.1
scikit-learn==1.5.2
xgboost==2.1.3
lightgbm==4.5.0
catboost==1.2.7
joblib==1.4.2

# ── Utils ──
pydantic==2.10.3
python-dotenv==1.0.1
